WHY I.F. (ISLAMIC FINANCE)
Al
Equal.green
We are different, but equal.
Y IF?
(Why Islamic Finance?)
0
9 02 .20
Al
Equal Collaborative Investment (FinTech) Software
as Base for Building
Islamic Finance Collaborative Investment and Project Management Environment.
Islamic Finance: Resilient, Fair, Intellectually Profound and Efficient Conceptual Framework
Islamic Finance is a profound conceptual framework of immutable
though dynamically adoptable to changing external
environment moral rules serving as a foundation for building and a system for running a proper, just and efficient
economy.
The applicable rules of the Islamic Finance are so basic and efficient, that they were laid dow
n into the operational
paradigm of the European financial system back in the early medieval period, with such instruments as international
cashless money/value transfers, guarantees, insurance, trust, aval, bill of exchange, check, partnership, agency and many
other solutions borrowed from the Arabic realm. It will not be an exaggeration to say that as the Arabic numerals and the
concept of algorithm had been wired into the deep structure of our computational methods, so Islamic Finance principles
were ori ginally weaved into the infrastructure of the global financial and economic systems.
Our philosophers, as religious so secular, have insistingly taught us since the days of conception of the intellectual though
t
and scientific thinking that absolute mora l principles do exist, and pursuing life based on such principles is not easy,
extremely demanding, but worthy and even the most important endeavor for the humans. And yet, somewhere in the past we
have chosen to fork our Western financial and economic sys tems from morality, probably, probably for utilitarian reasons
looking for the easier ways, just like we have erroneously chosen to divorce politics and morals, too. In a society, where
economy occupies overwhelming part of the living space, such initi ative was highly precarious. At certain stage it became
even unclear where the eligible domain of morality shall be located. First it was disbanded from finance, then from
economy, then from politics, then from daily routines. Where shall it dwell?
In Is
lamic worldview, moral principles are primary, they rule and drive the economic life. In Western view, which became
“conventional”, economy has its own life, and morality just recombines itself from the chaos of individual economic
interests, somehow, by i nvisible hand. Well…
Up to now, the experience of disbanding finance, economy and morality has not been encouraging.
The major arguments of the contemporary critics of the Islamic Finance approach is that the system it proposes is archaic
and so excessi vely complicated. This is extraordinarily strange argument, since the existing and actively developing
technologies we currently have been conceived and and being developing for nothing but solving the problems of
complexities arising in course of human ac tivities. Something which was too complicated to solve or manage in the past is
not so complicated to solve or manage now, thanks to the technological achievements. So, one cannot be convincing saying
that something complicated shall be abandoned only beca use it seems too complicated to implement.